Implement `@Hook` and create `RobotConfig.AutoInit`
New `Hook` allows for attaching methods to run on OpMode notifications with ease. This replaces the `@Cleanup`.

New RobotConfig option that allows users to use a `Hook` internally via a singleton for cleaner code and always-init. This is an optional feature and the old method will always continue to be supported for simplicity. Includes inhibition and better double-init protection.

`onInit()` methods made default implementations due to new RobotConfig options